remove user guide logic from content panel
