+++ apps/fishing-map/features/content-panel/UserGuideContent.tsx
@@ -4,24 +4,52 @@ import cx from 'classnames'
 
 import { Button, Icon } from '@globalfishingwatch/ui-components'
 
+import { fetchUserGuideContent } from 'features/cms/content.queries'
 import type { TUserGuideSection } from 'features/cms/strapi.types'
 import ContentHeader from 'features/content-panel/ContentHeader'
 import ContentMarkdown from 'features/content-panel/ContentMarkdown'
 import { useSidePanel } from 'features/content-panel/contentPanel.hooks'
+import EmptyContent from 'features/content-panel/EmptyContent'
 import TableOfContents from 'features/content-panel/TableOfContents'
 import { Route } from 'routes/_app'
+import type { Locale } from 'types'
 
 import styles from './ContentPanel.module.css'
 
-type UserGuideContentProps = { data: TUserGuideSection[] }
-
-export const UserGuideContent = ({ data }: UserGuideContentProps) => {
+export const UserGuideContent = () => {
+  const { status: loaderStatus, data: loaderData } = Route.useLoaderData()
   const { sidePanelId, sidePanelSubcontentId } = Route.useSearch()
+  const { i18n, t } = useTranslation()
+
+  const [langData, setLangData] = useState<{
+    status: string
+    data: TUserGuideSection[]
+  } | null>(null)
+
+  useEffect(() => {
+    const refetch = async (locale: Locale) => {
+      try {
+        const response = await fetchUserGuideContent({ locale })
+        if (!response?.data?.length) {
+          setLangData({ status: 'empty', data: [] })
+        } else {
+          setLangData({ status: 'success', data: response.data as TUserGuideSection[] })
+        }
+      } catch {
+        setLangData(null)
+      }
+    }
+    i18n.on('languageChanged', refetch)
+    return () => i18n.off('languageChanged', refetch)
+  }, [i18n])
+
+  const status = langData?.status ?? loaderStatus
+  const data = (langData?.data ?? loaderData) as TUserGuideSection[]
+
   const [isTableOfContentsOpen, setIsTableOfContentsOpen] = useState(!sidePanelId)
   const [isScrolled, setIsScrolled] = useState(false)
   const scrollContainerRef = useRef<HTMLDivElement>(null)
   const { openSidePanel } = useSidePanel()
-  const { t } = useTranslation()
 
   useEffect(() => {
     const el = scrollContainerRef.current
@@ -84,6 +112,8 @@ export const UserGuideContent = ({ data }: UserGuideContentProps) => {
       cancelled = true
     }
   }, [sidePanelSubcontentId, isTableOfContentsOpen, selectedSection])
+
+  if (status === 'error' || status === 'empty' || !data?.length) return <EmptyContent />
 
   return (
     <div className={cx(styles.container, { [styles.userGuideBackground]: !isTableOfContentsOpen })}>
